tpm2: Split up SealedKeyObject
SealedKeyObject will be deprecated by a follow up PR, and a new API
will be introduced to replace it. This splits it into code that can
be reused by the new API (sealedKeyDataBase) and moves the existing
API to a new file.
