+++ tpm2/keydata.go
@@ -21,34 +21,16 @@ package tpm2
 
 import (
 	"bytes"
-	"crypto"
 	"errors"
 	"fmt"
-	"hash"
 	"io"
-	"os"
 
 	"github.com/canonical/go-tpm2"
-	"github.com/canonical/go-tpm2/mu"
-	"github.com/snapcore/snapd/osutil"
-	"github.com/snapcore/snapd/osutil/sys"
 
 	"golang.org/x/xerrors"
 
-	"maze.io/x/crypto/afis"
-
-	"github.com/snapcore/secboot"
 	"github.com/snapcore/secboot/internal/tcg"
 )
-
-const (
-	keyDataHeader uint32 = 0x55534b24
-)
-
-type sealedData struct {
-	Key            secboot.DiskUnlockKey
-	AuthPrivateKey secboot.AuxiliaryKey
-}
 
 type keyDataError struct {
 	err error
@@ -137,20 +119,15 @@ func newKeyData(keyPrivate tpm2.Private, keyPublic *tpm2.Public, importSymSeed t
 	}
 }
 
-// SealedKeyObject corresponds to a sealed key data file.
-type SealedKeyObject struct {
+type sealedKeyDataBase struct {
 	data keyData
-}
-
-func newSealedKeyObject(data keyData) *SealedKeyObject {
-	return &SealedKeyObject{data: data}
 }
 
 // ensureImported will import the sealed key object into the TPM's storage hierarchy if
 // required, as indicated by an import symmetric seed of non-zero length. The tpmKeyData
 // structure will be updated with the newly imported private area and the import
 // symmetric seed will be cleared.
-func (k *SealedKeyObject) ensureImported(tpm *tpm2.TPMContext, parent tpm2.ResourceContext, session tpm2.SessionContext) error {
+func (k *sealedKeyDataBase) ensureImported(tpm *tpm2.TPMContext, parent tpm2.ResourceContext, session tpm2.SessionContext) error {
 	if len(k.data.ImportSymSeed()) == 0 {
 		return nil
 	}
@@ -166,7 +143,7 @@ func (k *SealedKeyObject) ensureImported(tpm *tpm2.TPMContext, parent tpm2.Resou
 
 // load loads the TPM sealed object associated with this keyData in to the storage hierarchy of the TPM, and returns the newly
 // created tpm2.ResourceContext.
-func (k *SealedKeyObject) load(tpm *tpm2.TPMContext, parent tpm2.ResourceContext, session tpm2.SessionContext) (tpm2.ResourceContext, error) {
+func (k *sealedKeyDataBase) load(tpm *tpm2.TPMContext, parent tpm2.ResourceContext, session tpm2.SessionContext) (tpm2.ResourceContext, error) {
 	if err := k.ensureImported(tpm, parent, session); err != nil {
 		return nil, err
 	}
@@ -175,7 +152,7 @@ func (k *SealedKeyObject) load(tpm *tpm2.TPMContext, parent tpm2.ResourceContext
 }
 
 // validateData performs correctness checks on this object.
-func (k *SealedKeyObject) validateData(tpm *tpm2.TPMContext, session tpm2.SessionContext) (*tpm2.NVPublic, error) {
+func (k *sealedKeyDataBase) validateData(tpm *tpm2.TPMContext, session tpm2.SessionContext) (*tpm2.NVPublic, error) {
 	sealedKeyTemplate := makeImportableSealedKeyTemplate()
 
 	// Perform some initial checks on the sealed data object's public area to
@@ -223,209 +200,4 @@ func (k *SealedKeyObject) validateData(tpm *tpm2.TPMContext, session tpm2.Sessio
 	}
 
 	return pcrPolicyCounterPub, nil
-}
-
-// Version returns the version number that this sealed key object was created with.
-func (k *SealedKeyObject) Version() uint32 {
-	return k.data.Version()
-}
-
-// PCRPolicyCounterHandle indicates the handle of the NV counter used for PCR policy revocation for this sealed key object (and for
-// PIN integration for version 0 key files).
-func (k *SealedKeyObject) PCRPolicyCounterHandle() tpm2.Handle {
-	return k.data.Policy().PCRPolicyCounterHandle()
-}
-
-// WriteAtomic will serialize this SealedKeyObject to the supplied writer.
-func (k *SealedKeyObject) WriteAtomic(w secboot.KeyDataWriter) error {
-	if _, err := mu.MarshalToWriter(w, k.data.Version()); err != nil {
-		return err
-	}
-	if err := k.data.Write(w); err != nil {
-		return err
-	}
-	return w.Commit()
-}
-
-// ReadSealedKeyObject reads a SealedKeyObject from the supplied io.Reader. If it
-// cannot be correctly decoded, an InvalidKeyDataError error will be returned.
-func ReadSealedKeyObject(r io.Reader) (*SealedKeyObject, error) {
-	var version uint32
-	if _, err := mu.UnmarshalFromReader(r, &version); err != nil {
-		return nil, InvalidKeyDataError{err.Error()}
-	}
-
-	data, err := readKeyData(r, version)
-	if err != nil {
-		return nil, InvalidKeyDataError{err.Error()}
-	}
-
-	return newSealedKeyObject(data), nil
-}
-
-type fileKeyDataHdr struct {
-	Magic   uint32
-	Version uint32
-}
-
-type stripedFileKeyDataHdr struct {
-	Stripes uint32
-	HashAlg tpm2.HashAlgorithmId
-	Size    uint32
-}
-
-// NewFileSealedKeyObjectReader creates an io.Reader from the file at the specified
-// path that can be passed to ReadSealedKeyObject. The file will have been previously
-// created by SealKeyToTPM. If the file cannot be opened, an *os.PathError error will
-// be returned.
-//
-// This function decodes part of the metadata specific to key files. If this fails,
-// an InvalidKeyDataError error will be returned.
-func NewFileSealedKeyObjectReader(path string) (io.Reader, error) {
-	f, err := os.Open(path)
-	if err != nil {
-		return nil, err
-	}
-	defer f.Close()
-
-	// v0 files contain the following structure:
-	//  magic   uint32 // 0x55534b24
-	//  version uint32 // 0
-	//  data    []byte
-	//
-	// post-v0 files contain the following structure:
-	//  magic	uint32 // 0x55534b24
-	//  version	uint32
-	//  stripes	uint32
-	//  hashAlg	tpm2.HashAlgorithmId
-	//  size	uint32
-	//  stripedData [size]byte
-	//
-	// We want to use the version field to encode the keyData structure
-	// version for all sources, but we only want to use the AF splitter
-	// for files. Ideally the key data version would be after the afis
-	// header, but it isn't. We do some manipulation here to move it so
-	// that the keyData unmarshaller can access it.
-
-	var hdr fileKeyDataHdr
-	if _, err := mu.UnmarshalFromReader(f, &hdr); err != nil {
-		return nil, InvalidKeyDataError{fmt.Sprintf("cannot unmarshal file header: %v", err)}
-	}
-
-	if hdr.Magic != keyDataHeader {
-		return nil, InvalidKeyDataError{fmt.Sprintf("unexpected magic (%d)", hdr.Magic)}
-	}
-
-	// Prepare a buffer for unmarshalling keyData.
-	buf := new(bytes.Buffer)
-	mu.MarshalToWriter(buf, hdr.Version)
-
-	if hdr.Version == 0 {
-		if _, err := io.Copy(buf, f); err != nil {
-			return nil, InvalidKeyDataError{fmt.Sprintf("cannot read data: %v", err)}
-		}
-		return buf, nil
-	}
-
-	var afisHdr stripedFileKeyDataHdr
-	if _, err := mu.UnmarshalFromReader(f, &afisHdr); err != nil {
-		return nil, InvalidKeyDataError{fmt.Sprintf("cannot unmarshal AFIS header: %v", err)}
-	}
-
-	if afisHdr.Stripes == 0 {
-		return nil, InvalidKeyDataError{"invalid number of stripes"}
-	}
-	if !afisHdr.HashAlg.Available() {
-		return nil, InvalidKeyDataError{"digest algorithm unavailable"}
-	}
-
-	data := make([]byte, afisHdr.Size)
-	if _, err := io.ReadFull(f, data); err != nil {
-		return nil, InvalidKeyDataError{fmt.Sprintf("cannot read striped data: %v", err)}
-	}
-
-	merged, err := afis.MergeHash(data, int(afisHdr.Stripes), func() hash.Hash { return afisHdr.HashAlg.NewHash() })
-	if err != nil {
-		return nil, InvalidKeyDataError{fmt.Sprintf("cannot merge data: %v", err)}
-	}
-
-	if _, err := buf.Write(merged); err != nil {
-		return nil, err
-	}
-
-	return buf, nil
-}
-
-type FileSealedKeyObjectWriter struct {
-	*bytes.Buffer
-	path string
-}
-
-func (w *FileSealedKeyObjectWriter) Commit() (err error) {
-	f, err := osutil.NewAtomicFile(w.path, 0600, 0, sys.UserID(osutil.NoChown), sys.GroupID(osutil.NoChown))
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err != nil {
-			f.Cancel()
-		} else {
-			err = f.Commit()
-		}
-	}()
-
-	hdr := fileKeyDataHdr{Magic: keyDataHeader}
-	if _, err := mu.UnmarshalFromReader(w, &hdr.Version); err != nil {
-		return err
-	}
-
-	if _, err := mu.MarshalToWriter(f, &hdr); err != nil {
-		return err
-	}
-
-	if hdr.Version == 0 {
-		if _, err := io.Copy(f, w); err != nil {
-			return err
-		}
-
-		return nil
-	}
-
-	stripes := uint32((128 * 1024 / w.Len()) + 1)
-
-	data, err := afis.SplitHash(w.Bytes(), int(stripes), func() hash.Hash { return crypto.SHA256.New() })
-	if err != nil {
-		return err
-	}
-
-	afisHdr := stripedFileKeyDataHdr{
-		Stripes: stripes,
-		HashAlg: tpm2.HashAlgorithmSHA256,
-		Size:    uint32(len(data))}
-	if _, err := mu.MarshalToWriter(f, &afisHdr); err != nil {
-		return err
-	}
-
-	if _, err := f.Write(data); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// NewFileSealedKeyObjectWriter creates a new writer for atomically updating a sealed key
-// data file using SealedKeyObject.WriteAtomic.
-func NewFileSealedKeyObjectWriter(path string) *FileSealedKeyObjectWriter {
-	return &FileSealedKeyObjectWriter{new(bytes.Buffer), path}
-}
-
-// ReadSealedKeyObjectFromFile reads a SealedKeyObject from the file created by SealKeyToTPM at the specified path.
-// If the file cannot be opened, an *os.PathError error is returned. If the file cannot be deserialized successfully,
-// an InvalidKeyDataError error will be returned.
-func ReadSealedKeyObjectFromFile(path string) (*SealedKeyObject, error) {
-	r, err := NewFileSealedKeyObjectReader(path)
-	if err != nil {
-		return nil, err
-	}
-	return ReadSealedKeyObject(r)
 }
